+++ src/index.ts
@@ -1,5 +1,11 @@
 import { Mutex } from '@livekit/mutex';
-import { DataPacket_Kind, DisconnectReason, SubscriptionError, TrackType } from '@livekit/protocol';
+import {
+  DataPacket_Kind,
+  DisconnectReason,
+  Encryption_Type,
+  SubscriptionError,
+  TrackType,
+} from '@livekit/protocol';
 import { LogLevel, LoggerNames, getLogger, setLogExtension, setLogLevel } from './logger';
 import DefaultReconnectPolicy from './room/DefaultReconnectPolicy';
 import type { ReconnectContext, ReconnectPolicy } from './room/ReconnectPolicy';
@@ -33,12 +39,14 @@ import {
   createAudioAnalyser,
   getEmptyAudioStreamTrack,
   getEmptyVideoStreamTrack,
+  isAudioCodec,
   isAudioTrack,
   isBrowserSupported,
   isLocalParticipant,
   isLocalTrack,
   isRemoteParticipant,
   isRemoteTrack,
+  isVideoCodec,
   isVideoTrack,
   supportsAV1,
   supportsAdaptiveStream,
@@ -58,6 +66,8 @@ export * from './room/errors';
 export * from './room/events';
 export * from './room/track/Track';
 export * from './room/track/create';
+export * from './room/token-source/TokenSource';
+export * from './room/token-source/types';
 export { facingModeFromDeviceLabel, facingModeFromLocalTrack } from './room/track/facingMode';
 export * from './room/track/options';
 export * from './room/track/processor/types';
@@ -79,6 +89,7 @@ export {
   ConnectionState,
   CriticalTimers,
   DataPacket_Kind,
+  Encryption_Type,
   DefaultReconnectPolicy,
   DisconnectReason,
   LocalAudioTrack,
@@ -113,9 +124,11 @@ export {
   supportsDynacast,
   supportsVP9,
   Mutex,
+  isAudioCodec,
   isAudioTrack,
   isLocalTrack,
   isRemoteTrack,
+  isVideoCodec,
   isVideoTrack,
   isLocalParticipant,
   isRemoteParticipant,